+++ nyquist/FLAC/src/libFLAC/crc.c
@@ -122,7 +122,7 @@ void FLAC__crc8_update_block(const FLAC__byte *data, unsigned len, FLAC__uint8 *
 		*crc = FLAC__crc8_table[*crc ^ *data++];
 }
 
-FLAC__uint8 FLAC__crc8(const FLAC__byte *data, unsigned len)
+FLAC__uint8 FLAC__crc8(const FLAC__byte *data, size_t len)
 {
 	FLAC__uint8 crc = 0;
 
@@ -132,7 +132,7 @@ FLAC__uint8 FLAC__crc8(const FLAC__byte *data, unsigned len)
 	return crc;
 }
 
-unsigned FLAC__crc16(const FLAC__byte *data, unsigned len)
+unsigned FLAC__crc16(const FLAC__byte *data, size_t len)
 {
 	unsigned crc = 0;
